Add feature to search with multiple pages
The bang query is now slightly different by making it a comma-delimited
list of bangs to be opened. Plugin activation event is also changed to
open the bangs from the field instead of the item, making search results
effective only for searching bangs and nothing else. This may change in
the future though if it proves to be inconvenient.