Add maximum limit for search results
This has an effect on perceived speed difference in searching. It is a
compromise especially that Pop shell launcher frontend can only display
8 items. Keeping the results small also gets more chances of getting
more relevant results especially the launcher service is combining the
results asynchronously. While there may be frontends that display more
than 8 items, we'll keep it small for now.