Add plugin-specific configuration
It's not yet complete and it is just a tentative feature. More options
are being considered and only the most basic things are in here for now.
I also decided the configuration to be a separate file instead of
extending `plugin.ron` since it isn't possible (at least from my
understanding of Rust and the codebase). Otherwise, it'll be transferred
there.