Rework the database
This shouldn't affect the end user. It is just for the sake of my
sanity. :)

There's also some refactoring of the codebase by putting them into the
more appropriate locations. The automatic handling of downloading
database now occurs when the app is initialized instead of putting it
when the database is loading. It is nice to make it more logically
sound.