Add `unique_bangs` option
Another update, another plugin option. This remove bangs with duplicate
URLs which is nice for removing aliases that the default database
apparently have like `youtube` being the same as `yt`, `ty`, and `you`
all of which are relevant and reaches the first few results. Though,
this option retains the most relevant so `yt` is the only retaining
bang.