Change the database loading mechanism
Instead of loading a database and saving it to a local file, we load it
directly through URL instead. This is better since you can also save
through `file://` scheme. Though, this is not good since it constantly
requests for the resource. We'll have to cache the resource with a time
limit on it later.